WIP: Use list item for entity picker selected item
